Update: fixed integer constraints and validations on reviews

+++ authenticate-me/backend/db/migrations/20230414221800-create-review.js
@@ -33,7 +33,7 @@ module.exports = {
         allowNull: false,
       },
       stars: {
-        type: Sequelize.INTEGER(1),// min 1 max 5
+        type: Sequelize.INTEGER,// min 1 max 5
         allowNull: false,
       },
       createdAt: {
